# crates/hkask-api/src/routes/chat.rs
//! Curator chat routes
//
//! The `POST /api/chat` endpoint accepts an optional `model` field that
//! switches the LLM used for inference. Use `GET /api/models` to discover
//! valid model identifiers.
//!
//! The `POST /api/chat/stream` endpoint streams inference output as SSE events.

use axum::extract::Extension;
use axum::{
    Json,
    extract::State,
    response::sse::{Event, Sse},
};
use std::convert::Infallible;
use utoipa_axum::{router::OpenApiRouter, routes};

use crate::ApiState;
use crate::middleware::auth::AuthContext;
use hkask_services::{ChatRequest as ServiceChatRequest, ChatService};
use hkask_types::LLMParameters;
use hkask_types::ports::InferencePort;
use serde::{Deserialize, Serialize};
use utoipa::ToSchema;

/// Chat request sent to the Curator or a specified agent.
///
/// The `model` field allows switching the LLM at request time. When omitted,
/// the server default (qwen3:8b) is used. Use `GET /api/models` to discover
/// available models, and `GET /api/models/search?q=...` for fuzzy matching.
#[derive(Debug, Serialize, Deserialize, ToSchema)]
pub struct ChatRequest {
    /// User input message
    pub input: String,
    /// Optional template ID to contextualize the prompt
    pub template_id: Option<String>,
    /// Model override for inference (e.g., "qwen3:8b"). If unset, uses the server default.
    #[serde(default)]
    pub model: Option<String>,
}

/// Chat response from the Curator or agent.
///
/// The `model` field echoes which LLM was used, confirming model switching.
#[derive(Debug, Serialize, Deserialize, ToSchema)]
pub struct ChatResponse {
    /// Generated response text
    pub output: String,
    /// Template ID that was applied (or "auto-select")
    pub template_id: String,
    /// Model identifier used for inference
    pub model: String,
}

/// Create chat router
pub fn chat_router() -> OpenApiRouter<ApiState> {
    OpenApiRouter::new()
        .routes(routes!(chat))
        .routes(routes!(chat_stream))
}

/// Chat with the Curator or a specified agent.
///
/// Accepts an optional `model` field to switch the LLM at request time.
/// When omitted, the server default (`qwen3:8b`) is used. The response
/// echoes the `model` used, confirming which LLM generated the output.
///
/// Use `GET /api/models` or `GET /api/models/search?q=...` to discover
/// available model identifiers.
#[utoipa::path(
    post,
    path = "/api/chat",
    tag = "chat",
    request_body = ChatRequest,
    responses(
        (status = 200, description = "Chat response", body = ChatResponse),
        (status = 400, description = "Invalid request"),
        (status = 500, description = "Internal server error"),
    ),
)]
pub(crate) async fn chat(
    State(state): State<ApiState>,
    Extension(auth): Extension<AuthContext>,
    Json(req): Json<ChatRequest>,
) -> Json<ChatResponse> {
    let model_str = req.model.clone().unwrap_or_else(|| "qwen3:8b".to_string());
    let model: &str = &model_str;
    let strategy = hkask_templates::PromptStrategy::from_input(&req.input);

    // Frame the prompt via PromptStrategy (API-specific — uses template_id)
    let prompt = match &req.template_id {
        Some(id) => format!("[template: {}] {}", id, req.input),
        None => strategy.frame(&req.input).to_string(),
    };

    let svc_req = ServiceChatRequest {
        input: prompt,
        agent_name: Some("Curator".to_string()),
        model_override: req.model,
        system_prompt_suffix: None,
        tool_section: None,
        inference_port_override: None,
        episodic_storage_override: None,
        semantic_storage_override: None,
        auth_context: Some(auth),
        params_override: None,
    };

    let result = match ChatService::chat(&state.agent_service, svc_req).await {
        Ok(resp) => resp,
        Err(e) => hkask_services::ChatResponse {
            text: format!("Chat error: {}", e),
            usage: None,
            finish_reason: "error".to_string(),
            tool_calls: vec![],
        },
    };

    Json(ChatResponse {
        output: result.text,
        template_id: req
            .template_id
            .unwrap_or_else(|| strategy.name().to_string()),
        model: model.to_string(),
    })
}

/// Stream chat inference output as Server-Sent Events.
///
/// Accepts the same `ChatRequest` as `/api/chat` but streams the inference
/// output as SSE `data:` events. Each event carries a JSON object with
/// `text_delta`, `model`, and optionally `finish_reason`. The final event
/// has `finish_reason: "stop"` and includes `usage` stats.
///
/// This is a surface-layer endpoint — it bypasses ChatService's full
/// pipeline (memory recall, episodic storage) and calls the inference port
/// directly for streaming. Use `/api/chat` for the full pipeline with
/// memory integration.
#[utoipa::path(
    post,
    path = "/api/chat/stream",
    tag = "chat",
    request_body = ChatRequest,
    responses(
        (status = 200, description = "SSE stream of chat chunks", content_type = "text/event-stream"),
        (status = 400, description = "Invalid request"),
        (status = 401, description = "Unauthorized"),
        (status = 500, description = "Internal server error"),
    ),
)]
pub(crate) async fn chat_stream(
    State(state): State<ApiState>,
    Extension(_auth): Extension<AuthContext>,
    Json(req): Json<ChatRequest>,
) -> Sse<impl futures_util::Stream<Item = Result<Event, Infallible>>> {
    let model_str = req.model.clone().unwrap_or_else(|| "qwen3:8b".to_string());
    let strategy = hkask_templates::PromptStrategy::from_input(&req.input);

    let prompt = match &req.template_id {
        Some(id) => format!("[template: {}] {}", id, req.input),
        None => strategy.frame(&req.input).to_string(),
    };

    let params = LLMParameters {
        temperature: 0.7,
        top_p: 0.9,
        top_k: 40,
        min_p: 0.0,
        typical_p: 0.0,
        frequency_penalty: 0.0,
        presence_penalty: 0.0,
        max_tokens: 512,
        seed: None,
    };

    let inference = state.agent_service.coordination().0.clone();

    // Use a channel to bridge the borrowed inference stream into a 'static
    // stream for the SSE response. A spawned task owns the Arc and sends
    // chunks through the channel.
    let (tx, rx): (tokio::sync::mpsc::Sender<Result<Event, Infallible>>, _) =
        tokio::sync::mpsc::channel(32);

    match inference {
        Some(port) => {
            tokio::spawn(async move {
                let mut stream =
                    port.generate_stream_with_model(&prompt, &params, Some(&model_str));
                use futures_util::StreamExt;
                while let Some(chunk_result) = stream.next().await {
                    let event = match chunk_result {
                        Ok(chunk) => {
                            let data = serde_json::json!({
                                "text_delta": chunk.text_delta,
                                "model": chunk.model,
                                "finish_reason": chunk.finish_reason,
                                "usage": chunk.usage.map(|u| serde_json::json!({
                                    "prompt_tokens": u.prompt_tokens,
                                    "completion_tokens": u.completion_tokens,
                                    "total_tokens": u.total_tokens,
                                })),
                                "tool_calls": if chunk.tool_calls.is_empty() { None } else { Some(serde_json::json!(chunk.tool_calls)) },
                            });
                            Event::default().data(data.to_string())
                        }
                        Err(e) => Event::default().data(
                            serde_json::json!({
                                "error": e.to_string(),
                            })
                            .to_string(),
                        ),
                    };
                    if tx.send(Ok(event)).await.is_err() {
                        break; // receiver dropped
                    }
                }
            });
        }
        None => {
            let _ = tx
                .send(Ok(Event::default().data(
                    serde_json::json!({
                        "error": "Inference unavailable — no model loaded",
                    })
                    .to_string(),
                )))
                .await;
        }
    }

    let stream = tokio_stream::wrappers::ReceiverStream::new(rx);
    Sse::new(Box::pin(stream)).keep_alive(axum::response::sse::KeepAlive::default())
}