# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hKask API",
    "description": "A Minimal Viable Container for Agents - HTTP API",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.27.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "hKask API server"
    }
  ],
  "paths": {
...
    "/api/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat with the Curator or a specified agent.",
        "description": "Accepts an optional `model` field to switch the LLM at request time.\nWhen omitted, the server default (`qwen3:8b`) is used. The response\nechoes the `model` used, confirming which LLM generated the output.\n\nUse `GET /api/models` or `GET /api/models/search?q=...` to discover\navailable model identifiers.",
        "operationId": "chat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Chat response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Stream chat inference output as Server-Sent Events.",
        "description": "Accepts the same `ChatRequest` as `/api/chat` but streams the inference\noutput as SSE `data:` events. Each event carries a JSON object with\n`text_delta`, `model`, and optionally `finish_reason`. The final event\nhas `finish_reason: \"stop\"` and includes `usage` stats.\n\nThis is a surface-layer endpoint — it bypasses ChatService's full\npipeline (memory recall, episodic storage) and calls the inference port\ndirectly for streaming. Use `/api/chat` for the full pipeline with\nmemory integration.",
        "operationId": "chat_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE stream of chat chunks",
            "content": {
              "text/event-stream": {}
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/cns/subscribe": {
      "get": {
        "tags": [
          "cns"
        ],
        "summary": "Subscribe to CNS events as an SSE stream.",
        "description": "The endpoint upgrades the HTTP response to a long-lived SSE connection.\nEvents matching the requested span namespaces are forwarded in real time.\nLag notifications are emitted when the client falls behind.",
        "operationId": "cns_subscribe",
        "parameters": [
          {
            "name": "spans",
            "in": "query",
            "description": "Span namespaces to subscribe to (e.g., [\"cns.tool\", \"cns.inference\"])",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE event stream",
            "content": {
              "text/event-stream": {}
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/api/cns/variety": {
      "get": {
        "tags": [
          "cns"
        ],
        "summary": "CNS variety endpoint",
        "operationId": "cns_variety",
        "responses": {
          "200": {
            "description": "CNS variety counters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CnsVarietyResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List all available models from Okapi.",
        "description": "Queries Okapi's `/api/tags` endpoint and returns metadata for each\nlocally loaded model. Returns an empty list if Okapi is unreachable\n(graceful degradation).",
        "operationId": "list_models",
        "responses": {
          "200": {
            "description": "List of available models",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            }
          },
          "503": {
            "description": "Okapi unreachable (returns empty list)"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "type": "object",
        "description": "Chat request sent to the Curator or a specified agent.\n\nThe `model` field allows switching the LLM at request time. When omitted,\nthe server default (qwen3:8b) is used. Use `GET /api/models` to discover\navailable models, and `GET /api/models/search?q=...` for fuzzy matching.",
        "required": [
          "input"
        ],
        "properties": {
          "input": {
            "type": "string",
            "description": "User input message"
          },
          "model": {
            "type": [
              "string",
              "null"
            ],
            "description": "Model override for inference (e.g., \"qwen3:8b\"). If unset, uses the server default."
          },
          "template_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional template ID to contextualize the prompt"
          }
        }
      },
      "ChatResponse": {
        "type": "object",
        "description": "Chat response from the Curator or agent.\n\nThe `model` field echoes which LLM was used, confirming model switching.",
        "required": [
          "output",
          "template_id",
          "model"
        ],
        "properties": {
          "model": {
            "type": "string",
            "description": "Model identifier used for inference"
          },
          "output": {
            "type": "string",
            "description": "Generated response text"
          },
          "template_id": {
            "type": "string",
            "description": "Template ID that was applied (or \"auto-select\")"
          }
        }
      },
...
      "CnsVarietyResponse": {
        "type": "object",
        "description": "CNS variety response",
        "required": [
          "domains",
          "total_deficit",
          "counters"
        ],
        "properties": {
          "counters": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/VarietyCounterResponse"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "total_deficit": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "ModelEntry": {
        "type": "object",
        "description": "A model available through Okapi.\n\nIncludes metadata from Okapi's `/api/tags` endpoint: model family,\nparameter count, quantization level, and disk size.",
        "required": [
          "name"
        ],
        "properties": {
          "family": {
            "type": [
              "string",
              "null"
            ],
            "description": "Model family (e.g., \"llama\", \"qwen2\")"
          },
          "name": {
            "type": "string",
            "description": "Model identifier (e.g., \"qwen3:8b\", \"llama3.1:70b\")"
          },
          "parameter_size": {
            "type": [
              "string",
              "null"
            ],
            "description": "Parameter count (e.g., \"8B\", \"70B\")"
          },
          "quantization_level": {
            "type": [
              "string",
              "null"
            ],
            "description": "Quantization level (e.g., \"Q4_0\", \"Q5_K_M\")"
          },
          "size_gb": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Model size in gigabytes"
          }
        }
      },
      "ModelListResponse": {
        "type": "object",
        "description": "Response containing available Okapi models.\n\nReturned by `GET /api/models` and `GET /api/models/search`.\nModel names from this response can be used as the `model` field\nin `POST /api/chat` requests.",
        "required": [
          "models",
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "description": "Total number of models in the list",
            "minimum": 0
          },
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelEntry"
            },
            "description": "List of available models"
          }
        }
      },
...
      "VarietyCounterResponse": {
        "type": "object",
        "description": "CNS variety counter response",
        "required": [
          "variety",
          "total",
          "entropy"
        ],
        "properties": {
          "entropy": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "variety": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      }
    }
  },
...
}